Split lib.rs: extract router.rs and middleware.rs
Move the router construction (ApiDoc, build_router, build_middleware_stack,
span factories, CORS layer, trace context extraction) into a new router
module. Move auth/user-agent middleware plus IP extraction helpers into
a new middleware module. Update budget.rs to reference the relocated
helpers via their new module paths (crate::middleware::extract_client_ip,
crate::util::error_response).

19 IP-extraction tests move to middleware.rs alongside the code they
cover; 6 budget/logging integration tests move to budget.rs where they
drive the budget middleware directly. lib.rs shrinks from ~1530 lines to
~170, containing only the public entry points (build_app, serve),
module declarations, and re-exports that preserve the existing public
API.
